- 在 DataPool 模型中添加 connectionState 字段，用于记录连接状态 - 实现 TcpClientManager 类，负责管理 TCP客户端连接 - 新增 DataIngestionService 和 ParsingRuleEngineService，用于处理数据入库和解析规则 - 实现 TcpClientDataSchedulerService，定时检查并调度 TCP 客户端数据池- 新增 TcpClientHandler，处理 Netty 客户端入站数据 - 更新 DataPoolManagementHandler，支持手动连接和断开 TCP 服务

+++ ruoyi-business/src/main/java/com/ruoyi/business/service/impl/DataPoolItemServiceImpl.java
@@ -375,22 +375,22 @@ public class DataPoolItemServiceImpl implements IDataPoolItemService {
         
         for (Map<String, Object> statusCount : statusCounts) {
             String status = (String) statusCount.get("status");
-            Long count = (Long) statusCount.get("count");
+            Integer count = (Integer) statusCount.get("count");
             
-            totalCount += count.intValue();
+            totalCount += count;
             
             switch (status) {
                 case "PENDING":
-                    pendingCount = count.intValue();
+                    pendingCount = count;
                     break;
                 case "PRINTING":
-                    printingCount = count.intValue();
+                    printingCount = count;
                     break;
                 case "PRINTED":
-                    printedCount = count.intValue();
+                    printedCount = count;
                     break;
                 case "FAILED":
-                    failedCount = count.intValue();
+                    failedCount = count;
                     break;
             }
         }
